Checkpoint: 阶段1-1分析完成，准备激进重构
完成工作：
- Track A: Process类型分析（4个定义）
- Track B: 临时代码识别（70+文件）
- Track C: 大文件识别（8个文件）
- Track D: 内存管理分析（13个分配器）
- Track E: 目录结构分析（3个循环依赖）

生成文档：
- 15个详细报告和脚本
- 综合总结文档
- 实施方案

下一步：并行执行所有Track的重构工作

+++ kernel/src/posix/types.rs
@@ -2,8 +2,11 @@
 //!
 //! Standard primitive types for POSIX compliance
 
+// Allow POSIX-style type names (snake_case with _t suffix) which are standard in POSIX
+#![allow(non_camel_case_types)]
+
 /// Process ID type
-pub type Pid = usize;
+pub type Pid = i32;
 
 /// User ID type
 pub type Uid = u32;
